refactor(checker): split indexed access object formatting
