+++ crates/tsz-checker/src/types/type_checking/indexed_access.rs
@@ -6,6 +6,7 @@ use tsz_solver::TypeId;
 
 mod indexed_access_helpers;
 mod mapped_key_check;
+mod object_format;
 
 use indexed_access_helpers::{
     generic_constrained_index, indexed_access_object_alias_application_exceeds_depth,
@@ -62,25 +63,6 @@ impl<'a> CheckerState<'a> {
             }
             _ => false,
         }
-    }
-
-    fn format_ts2536_object_type(&self, object_node_idx: NodeIndex, object_type: TypeId) -> String {
-        if let Some(node) = self.ctx.arena.get(object_node_idx)
-            && matches!(
-                node.kind,
-                k if k == syntax_kind_ext::TYPE_REFERENCE
-                    || k == syntax_kind_ext::INDEXED_ACCESS_TYPE
-            )
-            && let Some(text) = self.node_text(object_node_idx)
-        {
-            let text = text.trim();
-            let text = text.strip_prefix('(').unwrap_or(text);
-            let text = text.strip_suffix(')').unwrap_or(text).trim();
-            if !text.is_empty() {
-                return text.to_string();
-            }
-        }
-        self.format_type(object_type)
     }
 
     fn typeof_global_this_indexed_key_is_missing(&self, key: &str) -> bool {
